moar logging; change test studies

+++ otindex/add_update_studies.py
@@ -28,7 +28,7 @@ def add_study(study_id):
     try:
         studyobj = phy.get_study(study_id)['data']
     except:
-        _LOG.debug('did not find study {s}'.format(s=study_id))
+        _LOG.debug('did not find study {s} in phylesystem'.format(s=study_id))
         raise HTTPNotFound("Study {s} not found in phylesystem".format(s=study_id))
     nexml = get_nexml_el(studyobj)
     proposedTrees = nexml.get('^ot:candidateTreeForSynthesis')
@@ -158,27 +158,24 @@ def deleteOrphanedCurators(study_id):
         )
         # if there is only one study edited by this curator
         if (studies.count()==1):
-            print "deleting curator {i}".format(i=curator_id)
+            _LOG.debug("deleting orphan curator {i} for study {s}".format(i=curator_id,s=study_id))
             DBSession.delete(
                 DBSession.query(Curator).filter(
                     Curator.id==curator_id
                 ).one()
             )
+        else:
+            _LOG.debug("no orphaned curators for study {s}".format(s=study_id))
 
+
+# should only be called after determining that study exists
 def delete_study(study_id):
     study = DBSession.query(Study).filter(
         Study.id == study_id
     ).first()
-    if (study):
-        _LOG.debug("deleting study {s}".format(s=study_id))
-        # check for to-be-orphaned curators
-        deleteOrphanedCurators(study_id)
-        DBSession.delete(study)
-        # need to flush here before re-adding updated study
-        DBSession.flush()
-    else:
-        _LOG.debug("did not find study {s} to delete".format(s=study_id))
-        raise HTTPNotFound("study id {s} not found".format(s=study_id))
+    # check for to-be-orphaned curators
+    deleteOrphanedCurators(study_id)
+    DBSession.delete(study)
 
 # get the lineage from this ID back to the root node
 def get_lineage(ott_id):
@@ -223,17 +220,20 @@ def get_study_id_from_url(url):
 
 # called directly by view add_update_studies_v3
 def remove_study(studyid):
+    study_id = studyid
     try:
-        study_id = studyid
         if studyid.startswith('http'):
             study_id = get_study_id_from_url(studyid)
-        _LOG.debug("removing study {s}".format(s=study_id))
-        delete_study(study_id)
     except HTTPNotFound:
         raise
-    except Exception:
-        raise
+    if study_exists(study_id):
+        _LOG.debug("deleting study {s}".format(s=study_id))
+        delete_study(study_id)
+    else:
+        _LOG.debug("did not find study {s} in database; cannot delete".format(s=study_id))
+        raise HTTPNotFound("study {s} not found in database".format(s=study_id))
 
+# check if study exists in DB
 def study_exists(study_id):
     study = DBSession.query(Study.id).filter(Study.id==study_id).first()
     if study:
@@ -247,9 +247,11 @@ def update_study(studyid):
         study_id = studyid
         if studyid.startswith('http'):
             study_id = get_study_id_from_url(studyid)
-        _LOG.debug("adding / updating study {s}".format(s=study_id))
+        # delete if already exists
         if (study_exists(study_id)):
             delete_study(study_id)
+            # need to flush here before re-adding updated study
+            DBSession.flush()
         #with DBSession.no_autoflush:
         add_study(study_id)
     except Exception:
